migration: emailAuthCode 컨트롤러, 서비스, 레포

+++ backend/src/main/java/com/backend/petplace/domain/user/service/UserService.java
@@ -44,8 +44,11 @@ public class UserService {
   @Transactional
   protected void checkAuthCode(UserSignupRequest request) {
     EmailAuthCode emailAuthCode = emailAuthCodeRepository.findByEmailAndAuthCode
-            (request.getEmail(), request.getAuthCode())
-        .orElseThrow(() -> new BusinessException(ErrorCode.AUTH_CODE_NOT_FOUND));
+            (request.getEmail(), request.getAuthCode());
+
+    if (emailAuthCode == null) {
+      throw new BusinessException(ErrorCode.AUTH_CODE_NOT_FOUND);
+    }
 
     if (emailAuthCode.getVerified()) {
       emailAuthCodeRepository.delete(emailAuthCode);
